Fixed format expression

+++ include/buffer.h
@@ -162,7 +162,7 @@ namespace pipes {
 			inline N_T at(size_t index) const {
 				if(this->length() <= index) {
                     char buffer[256];
-					print_formated(buffer, 256, "Index %lu is out of range. Max allowed %lu", index, this->length());
+					print_formated(buffer, 256, "Index %zu is out of range. Max allowed %zu", (size_t) index, (size_t) this->length());
                     throw std::out_of_range(buffer);
                 }
 				return (N_T) *(T*) (this->data_ptr<char>() + index);
@@ -173,7 +173,7 @@ namespace pipes {
 			inline const T& at(size_t index) const {
 				if(this->length() <= index) {
                     char buffer[256];
-					print_formated(buffer, 256, "Index %lu is out of range. Max allowed %lu", index, this->length());
+					print_formated(buffer, 256, "Index %zu is out of range. Max allowed %zu", (size_t) index, (size_t) this->length());
                     throw std::out_of_range(buffer);
                 }
 				return *(T*) (this->data_ptr<char>() + index);
